Move getSingleWord into a separate function
